+++ src/main/java/org/ancestra/evolutive/game/packet/guild/RemoveCollector.java
@@ -32,7 +32,7 @@ public class RemoveCollector implements PacketParser {
 				String str = "";
 				str += "R"+perco.get_N1()+","+perco.get_N2()+"|";
 				str += perco.get_mapID()+"|";
-				str += World.data.getCarte((short)perco.get_mapID()).getX()+"|"+World.data.getCarte((short)perco.get_mapID()).getY()+"|"+client.getPlayer().getName();
+				str += World.data.getCarte(perco.get_mapID()).getX()+"|"+World.data.getCarte(perco.get_mapID()).getY()+"|"+client.getPlayer().getName();
 				SocketManager.GAME_SEND_gT_PACKET(z, str);
 			}
 		}
